Conflicts:
	framework/phresco-framework-impl/src/main/java/com/photon/phresco/framework/impl/ProjectRuntimeManagerImpl.java

+++ plugins/java-maven-plugin/src/main/java/com/photon/phresco/plugins/JavaDeploy.java
@@ -19,9 +19,11 @@
  */
 package com.photon.phresco.plugins;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FilenameFilter;
 import java.io.IOException;
+import java.io.InputStreamReader;
 import java.util.List;
 
 import org.apache.maven.plugin.AbstractMojo;
@@ -90,7 +92,6 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 
 	public void execute() throws MojoExecutionException {
 		init();
-		updateFinalName();
 		createDb();
 		extractBuild();
 		deployToServer();
@@ -120,34 +121,6 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 						+ " -DenvironmentName=\"Multivalued evnironment names\""
 						+ " -DimportSql=\"Does the deployment needs to import sql(TRUE/FALSE?)\"");
 		throw new MojoExecutionException("Invalid Usage. Please see the Usage of Deploy Goal");
-	}
-
-	private void updateFinalName() throws MojoExecutionException {
-
-		try {
-				ProjectAdministrator projAdmin = PhrescoFrameworkFactory.getProjectAdministrator();
-				String envName = environmentName;
-				if (environmentName.indexOf(',') > -1) { // multi-value
-					envName = projAdmin.getDefaultEnvName(baseDir.getName());
-				}
-				List<SettingsInfo> settingsInfos = projAdmin.getSettingsInfos(Constants.SETTINGS_TEMPLATE_SERVER,
-						baseDir.getName(), envName);
-				for (SettingsInfo settingsInfo : settingsInfos) {
-					context = settingsInfo.getPropertyInfo(Constants.SERVER_CONTEXT).getValue();
-					break;
-				}
-			File pom = project.getFile();
-			PomProcessor pomprocessor = new PomProcessor(pom);
-			pomprocessor.setFinalName(context);
-			pomprocessor.save();
-
-		} catch (IOException e) {
-			throw new MojoExecutionException(e.getMessage(), e);
-		} catch (PhrescoException e) {
-			throw new MojoExecutionException(e.getMessage(), e);
-		} catch (Exception e) {
-			throw new MojoExecutionException(e.getMessage(), e);
-		}
 	}
 
 	private void createDb() throws MojoExecutionException {
@@ -190,15 +163,16 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 		}
 		String serverhost = info.getPropertyInfo(Constants.SERVER_HOST).getValue();
 		String serverport = info.getPropertyInfo(Constants.SERVER_PORT).getValue();
+		String serverprotocol = info.getPropertyInfo(Constants.SERVER_PROTOCOL).getValue();
 		String serverusername = info.getPropertyInfo(Constants.SERVER_ADMIN_USERNAME).getValue();
 		String serverpassword = info.getPropertyInfo(Constants.SERVER_ADMIN_PASSWORD).getValue();
 		String version = info.getPropertyInfo(Constants.SERVER_VERSION).getValue();
 		String servertype = info.getPropertyInfo(Constants.SERVER_TYPE).getValue();
-		String context = info.getPropertyInfo(Constants.SERVER_CONTEXT).getValue();
-
+		context = info.getPropertyInfo(Constants.SERVER_CONTEXT).getValue();
+		renameWar(context);
 		// no remote deployment
 		if (serverusername.isEmpty() && serverpassword.isEmpty()) {
-			renameWar(context);
+		//	renameWar(context);
 			deploy();
 			return;
 		}
@@ -206,11 +180,11 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 		// remote deployment
 		if (servertype.contains(TYPE_TOMCAT)
 				&& ((version.equals("7.0.x")) || (version.equals("7.1.x")) || (version.equals("6.0.x")))) {
-			deployToTomcatServer(serverhost, serverport, serverusername, serverpassword);
+			deployToTomcatServer(serverprotocol, serverhost, serverport, serverusername, serverpassword);
 		} else if (servertype.contains(TYPE_JBOSS) && (version.equals("7.0.x"))) {
-			deployToJbossServer(serverhost, serverusername, serverpassword);
+			deployToJbossServer(serverport, serverprotocol, serverhost, serverusername, serverpassword);
 		} else if (servertype.contains(TYPE_WEBLOGIC) && (version.equals("12c(12.1.1)"))) {
-			deployToWeblogicServer(serverhost, serverport, serverusername, serverpassword);
+			deployToWeblogicServer(serverprotocol, serverhost, serverport, serverusername, serverpassword);
 		} else {
 			// for other servers
 			deploy();
@@ -232,10 +206,11 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 	}
 	
 	
-	private void deployToTomcatServer(String serverhost, String serverport, String serverusername,
-			String serverpassword) throws MojoExecutionException {
+	private void deployToTomcatServer(String serverprotocol, String serverhost, String serverport,
+			String serverusername, String serverpassword) throws MojoExecutionException {
+		BufferedReader in = null;
+		boolean errorParam = false;
 		try {
-			getLog().info("project is deploying ......");
 			StringBuilder sb = new StringBuilder();
 			sb.append(MVN_CMD);
 			sb.append(STR_SPACE);
@@ -254,19 +229,36 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 			sb.append(serverpassword);
 			sb.append(STR_SPACE);
 			sb.append(SKIP_TESTS);
-			
 			Commandline cl = new Commandline(sb.toString());
+			Process process = cl.execute();
 			cl.setWorkingDirectory(baseDir);
-				Process process = cl.execute();
-			} catch (CommandLineException e) {
-				throw new MojoExecutionException(e.getMessage(), e);
+			in = new BufferedReader(new InputStreamReader(process.getInputStream()));
+			String line = null;
+			while ((line = in.readLine()) != null) {
+				if (line.startsWith("[ERROR]")) {
+					System.out.println(line); //do not use getLog() here as this line already contains the log type.
+					errorParam = true;
+				}
 			}
+			if (errorParam) {
+				throw new MojoExecutionException("Remote Deployment Failed ");
+			} else {
+				getLog().info(
+						" Project is Deploying into " + serverprotocol + "://" + serverhost + ":" + serverport + "/"
+								+ context);
+			}
+		} catch (CommandLineException e) {
+			throw new MojoExecutionException(e.getMessage(), e);
+		} catch (IOException e) {
+			throw new MojoExecutionException(e.getMessage(), e);
+		}
 	}
 
-	private void deployToJbossServer(String serverhost, String serverusername, String serverpassword)
+	private void deployToJbossServer(String serverport, String serverprotocol, String serverhost, String serverusername, String serverpassword)
 			throws MojoExecutionException {
+		BufferedReader in = null;
+		boolean errorParam = false;
 		try {
-			getLog().info("project is deploying ......");
 			StringBuilder sb = new StringBuilder();
 			sb.append(MVN_CMD);
 			sb.append(STR_SPACE);
@@ -285,16 +277,34 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 			
 			Commandline cl = new Commandline(sb.toString());
 			cl.setWorkingDirectory(baseDir);
-				Process process = cl.execute();
+			Process process = cl.execute();
+			in = new BufferedReader(new InputStreamReader(process.getInputStream()));
+			String line = null;
+			while ((line = in.readLine()) != null) {
+				if (line.startsWith("[ERROR]")) {
+					System.out.println(line); //do not use getLog() here as this line already contains the log type.
+					errorParam = true;
+				}
+			}
+			if (errorParam) {
+				throw new MojoExecutionException("Remote Deployment Failed ");
+			} else {
+				getLog().info(
+						" Project is Deploying into " + serverprotocol + "://" + serverhost + ":" + serverport + "/"
+								+ context);
+			}
 			} catch (CommandLineException e) {
+				throw new MojoExecutionException(e.getMessage(), e);
+			} catch (IOException e) {
 				throw new MojoExecutionException(e.getMessage(), e);
 			}
 	}
 
-	private void deployToWeblogicServer(String serverhost, String serverport, String serverusername,
+	private void deployToWeblogicServer(String serverprotocol, String serverhost, String serverport, String serverusername,
 			String serverpassword) throws MojoExecutionException {
+		BufferedReader in = null;
+		boolean errorParam = false;
 		try {
-			getLog().info("project is deploying ......");
 			StringBuilder sb = new StringBuilder();
 			sb.append(MVN_CMD);
 			sb.append(STR_SPACE);
@@ -317,7 +327,24 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 			Commandline cl = new Commandline(sb.toString());
 			cl.setWorkingDirectory(baseDir);
 				Process process = cl.execute();
+				in = new BufferedReader(new InputStreamReader(process.getInputStream()));
+				String line = null;
+				while ((line = in.readLine()) != null) {
+					if (line.startsWith("[ERROR]")) {
+						System.out.println(line); //do not use getLog() here as this line already contains the log type.
+						errorParam = true;
+					}
+				}
+				if (errorParam) {
+					throw new MojoExecutionException("Remote Deployment Failed ");
+				} else {
+					getLog().info(
+							" Project is Deploying into " + serverprotocol + "://" + serverhost + ":" + serverport + "/"
+									+ context);
+				}
 			} catch (CommandLineException e) {
+				throw new MojoExecutionException(e.getMessage(), e);
+			} catch (IOException e) {
 				throw new MojoExecutionException(e.getMessage(), e);
 			}
 	}
@@ -349,8 +376,10 @@ public class JavaDeploy extends AbstractMojo implements PluginConstants {
 				break;
 			}		
 			File deployDir = new File(deployLocation);
-			FileUtils.mkdir(deployDir.getPath().trim());	
-			getLog().info("Project is deploying.........");
+			if (!deployDir.exists()) {
+				throw new MojoExecutionException(" Deploy Directory" + deployLocation + " Does Not Exists ");
+			}
+			getLog().info("Project is deploying into " + deployLocation);
 			FileUtils.copyDirectoryStructure(tempDir.getAbsoluteFile(), deployDir);
 			getLog().info("Project is deployed successfully");
 		} catch (Exception e) {
